devops/requirements.txt
Flask>=3.0
redis>=5.0
psycopg2-binary>=2.9
